added auth and user

+++ controllers/posts.js
@@ -1,8 +1,17 @@
-const { Post, Comment } = require("../models");
+const { User, Post, Comment } = require("../models");
 
-const getAllPosts = async (_req, res) => {
+const getAllPosts = async (req, res) => {
+  const {
+    query: { page = 1, limit = 10 },
+  } = req;
+  const queryPage = Number(page),
+    queryLimit = Number(limit);
+
   try {
-    const posts = await Post.findAll();
+    const posts = await Post.findAll({
+      limit: queryLimit,
+      offset: (queryPage - 1) * queryLimit,
+    });
     return res.status(200).json(posts);
   } catch (error) {
     console.log("Error from get posts : ", error);
@@ -11,22 +20,35 @@ const getAllPosts = async (_req, res) => {
 };
 
 const createPost = async (req, res) => {
-  const post = req.body;
+  const { body, username } = req;
   try {
-    const data = await Post.create(post);
+    const data = await Post.create({ ...body, username });
     return res.status(201).json(data);
   } catch (error) {
     console.log("Error from create post : ", error);
-    return res.json(error);
+    return res.status(400).json(error);
   }
 };
 
 const getPostById = async (req, res) => {
-  const id = req.params.id;
+  const {
+    params: { id },
+    query: { page = 1, limit = 10 },
+  } = req;
+  const queryPage = Number(page),
+    queryLimit = Number(limit);
+
   try {
     const post = await Post.findOne({
       where: { id },
-      include: [{ model: Comment, as: "comments" }],
+      include: [
+        {
+          model: Comment,
+          as: "comments",
+          limit: queryLimit,
+          offset: (queryPage - 1) * queryLimit,
+        },
+      ],
     });
     if (!post) {
       return res.status(404).json({ message: "Post not found!" });
@@ -39,11 +61,26 @@ const getPostById = async (req, res) => {
 };
 
 const getPostsByUserName = async (req, res) => {
-  const username = req.params.username;
+  const {
+    params: { username },
+    query: { page = 1, limit = 10 },
+  } = req;
+  const queryPage = Number(page),
+    queryLimit = Number(limit);
+
   try {
-    const post = await Post.findAll({ where: { username } });
+    const user = await User.findOne({ where: { username } });
+    if (!user) {
+      return res.status(404).json({ error: "User not found!" });
+    }
+
+    const post = await Post.findAll({
+      where: { username },
+      limit: queryLimit,
+      offset: (queryPage - 1) * queryLimit,
+    });
     if (!post) {
-      return res.status(404).json({ message: "Post not found!" });
+      return res.status(404).json({ error: "Post not found!" });
     }
     return res.status(200).json(post);
   } catch (error) {
@@ -55,10 +92,10 @@ const getPostsByUserName = async (req, res) => {
 const deletePostById = async (req, res) => {
   const id = req.params.id;
   try {
-    const res = await Post.destroy({ where: { id } });
-    if (res !== 1) {
+    const count = await Post.destroy({ where: { id } });
+    if (count !== 1) {
       return res.status(404).json({
-        message: `Post not found!`,
+        error: "Post not found!",
       });
     }
     return res.status(204).json();
@@ -71,10 +108,15 @@ const deletePostById = async (req, res) => {
 const deletePostsByUserName = async (req, res) => {
   const username = req.params.username;
   try {
-    const res = await Post.destroy({ where: { username } });
-    if (res < 1) {
+    const user = await User.findOne({ where: { username } });
+    if (!user) {
+      return res.status(404).json({ error: "User not found!" });
+    }
+
+    const count = await Post.destroy({ where: { username } });
+    if (count < 1) {
       return res.status(404).json({
-        message: `Post not found!`,
+        message: "Post not found!",
       });
     }
     return res.status(204).json();
